Refactor package.json and package-lock.json to add pdf-lib dependency Remove unused import and commented out code in index.js
