+++ Backend/assets/dummy.js
@@ -1,0 +1,469 @@
+const dummydata = {
+
+    "semester": "First Semester",
+    "year": "2015-16",
+    "gender": "all",
+    "course_Observations": {
+        "GOOD": [
+            {
+                "course_code": "232 PDS-2",
+                "course_name": "Removable Prosthodontics (Preclinical II)",
+                "gender": "female",
+                "_id": "66e62b9afaf568e0518c7fa6"
+            },
+            {
+                "course_code": "121 RDS-1",
+                "course_name": "Dental Biomaterials I",
+                "gender": "male",
+                "_id": "66e62b9afaf568e0518c7fa7"
+            }
+        ],
+        "AVERAGE": [],
+        "POOR": [
+            {
+                "course_code": "232 PDS-2",
+                "course_name": "Removable Prosthodontics (Preclinical II)",
+                "gender": "male",
+                "_id": "66e62b9afaf568e0518c7fa8"
+            },
+            {
+                "course_code": "415 VDS-3",
+                "course_name": "Orthodontics I",
+                "gender": "female",
+                "_id": "66e62b9afaf568e0518c7fa9"
+            },
+            {
+                "course_code": "412 VDS-2",
+                "course_name": "Periodontology I",
+                "gender": "male",
+                "_id": "66e62b9afaf568e0518c7faa"
+            },
+            {
+                "course_code": "512 VDS-2",
+                "course_name": "Dental Public Health & Community Dentistry II",
+                "gender": "male",
+                "_id": "66e62b9afaf568e0518c7fab"
+            }
+        ]
+    },
+    "levelTable": [
+        {
+            "level": "6",
+            "classId": [
+                "66e4f8fe28bff2be4ed966f6",
+                "66e4f94d28bff2be4ed986fe"
+            ],
+            "levelAverage": {
+                "Poor (Bad) Questions": {
+                    "number": 34,
+                    "percentage": 42.75
+                },
+                "Very Difficult Question": {
+                    "number": 3,
+                    "percentage": 3.875
+                },
+                "Difficult Question": {
+                    "number": 4,
+                    "percentage": 5
+                },
+                "Good Question": {
+                    "number": 17,
+                    "percentage": 21.125
+                },
+                "Easy Question": {
+                    "number": 14,
+                    "percentage": 17.75
+                },
+                "Very Easy Question": {
+                    "number": 8,
+                    "percentage": 10
+                },
+                "Total Accepted": {
+                    "number": 17.5,
+                    "percentage": 21.875
+                },
+                "Total Rejected": {
+                    "number": 22.5,
+                    "percentage": 28.125
+                }
+            },
+            "_id": "66e62b9afaf568e0518c7fac"
+        },
+        {
+            "level": "3",
+            "classId": [
+                "66e5fed6d0f928ccf4bbbbd0"
+            ],
+            "levelAverage": {
+                "Poor (Bad) Questions": {
+                    "number": 34,
+                    "percentage": 43
+                },
+                "Very Difficult Question": {
+                    "number": 3,
+                    "percentage": 4
+                },
+                "Difficult Question": {
+                    "number": 4,
+                    "percentage": 5
+                },
+                "Good Question": {
+                    "number": 17,
+                    "percentage": 21
+                },
+                "Easy Question": {
+                    "number": 14,
+                    "percentage": 18
+                },
+                "Very Easy Question": {
+                    "number": 8,
+                    "percentage": 10
+                },
+                "Total Accepted": {
+                    "number": 0,
+                    "percentage": 0
+                },
+                "Total Rejected": {
+                    "number": 0,
+                    "percentage": 0
+                }
+            },
+            "_id": "66e62b9afaf568e0518c7fad"
+        },
+        {
+            "level": "10",
+            "classId": [
+                "66e60a15055142a83e52037e"
+            ],
+            "levelAverage": {
+                "Poor (Bad) Questions": {
+                    "number": 29,
+                    "percentage": 48
+                },
+                "Very Difficult Question": {
+                    "number": 4,
+                    "percentage": 7
+                },
+                "Difficult Question": {
+                    "number": 0,
+                    "percentage": 0
+                },
+                "Good Question": {
+                    "number": 9,
+                    "percentage": 15
+                },
+                "Easy Question": {
+                    "number": 3,
+                    "percentage": 5
+                },
+                "Very Easy Question": {
+                    "number": 15,
+                    "percentage": 25
+                },
+                "Total Accepted": {
+                    "number": 0,
+                    "percentage": 0
+                },
+                "Total Rejected": {
+                    "number": 0,
+                    "percentage": 0
+                }
+            },
+            "_id": "66e62b9afaf568e0518c7fae"
+        },
+        {
+            "level": "2",
+            "classId": [
+                "66e60ac858d7ef3fc302bd87"
+            ],
+            "levelAverage": {
+                "Poor (Bad) Questions": {
+                    "number": 29,
+                    "percentage": 48
+                },
+                "Very Difficult Question": {
+                    "number": 4,
+                    "percentage": 7
+                },
+                "Difficult Question": {
+                    "number": 0,
+                    "percentage": 0
+                },
+                "Good Question": {
+                    "number": 9,
+                    "percentage": 15
+                },
+                "Easy Question": {
+                    "number": 3,
+                    "percentage": 5
+                },
+                "Very Easy Question": {
+                    "number": 15,
+                    "percentage": 25
+                },
+                "Total Accepted": {
+                    "number": 0,
+                    "percentage": 0
+                },
+                "Total Rejected": {
+                    "number": 0,
+                    "percentage": 0
+                }
+            },
+            "_id": "66e62b9afaf568e0518c7faf"
+        },
+        {
+            "level": "12",
+            "classId": [
+                "66e60b5a2287e7f0cca09b2b"
+            ],
+            "levelAverage": {
+                "Poor (Bad) Questions": {
+                    "number": 44,
+                    "percentage": 73
+                },
+                "Very Difficult Question": {
+                    "number": 0,
+                    "percentage": 0
+                },
+                "Difficult Question": {
+                    "number": 2,
+                    "percentage": 3
+                },
+                "Good Question": {
+                    "number": 7,
+                    "percentage": 12
+                },
+                "Easy Question": {
+                    "number": 3,
+                    "percentage": 5
+                },
+                "Very Easy Question": {
+                    "number": 4,
+                    "percentage": 7
+                },
+                "Total Accepted": {
+                    "number": 0,
+                    "percentage": 0
+                },
+                "Total Rejected": {
+                    "number": 0,
+                    "percentage": 0
+                }
+            },
+            "_id": "66e62b9afaf568e0518c7fb0"
+        }
+    ],
+    "CourseNameTable": [
+        {
+            "CourseName": "Removable Prosthodontics (Preclinical II)",
+            "classId": [
+                "66e4f8fe28bff2be4ed966f6",
+                "66e4f94d28bff2be4ed986fe"
+            ],
+            "levelAverage": {
+                "Poor (Bad) Questions": {
+                    "number": 0,
+                    "percentage": 0
+                },
+                "Very Difficult Question": {
+                    "number": 0,
+                    "percentage": 0
+                },
+                "Difficult Question": {
+                    "number": 0,
+                    "percentage": 0
+                },
+                "Good Question": {
+                    "number": 0,
+                    "percentage": 0
+                },
+                "Easy Question": {
+                    "number": 0,
+                    "percentage": 0
+                },
+                "Very Easy Question": {
+                    "number": 0,
+                    "percentage": 0
+                },
+                "Total Accepted": {
+                    "number": 0,
+                    "percentage": 0
+                },
+                "Total Rejected": {
+                    "number": 0,
+                    "percentage": 0
+                }
+            },
+            "_id": "66e62b9afaf568e0518c7fb1"
+        },
+        {
+            "CourseName": "Dental Biomaterials I",
+            "classId": [
+                "66e5fed6d0f928ccf4bbbbd0"
+            ],
+            "levelAverage": {
+                "Poor (Bad) Questions": {
+                    "number": 0,
+                    "percentage": 0
+                },
+                "Very Difficult Question": {
+                    "number": 0,
+                    "percentage": 0
+                },
+                "Difficult Question": {
+                    "number": 0,
+                    "percentage": 0
+                },
+                "Good Question": {
+                    "number": 0,
+                    "percentage": 0
+                },
+                "Easy Question": {
+                    "number": 0,
+                    "percentage": 0
+                },
+                "Very Easy Question": {
+                    "number": 0,
+                    "percentage": 0
+                },
+                "Total Accepted": {
+                    "number": 0,
+                    "percentage": 0
+                },
+                "Total Rejected": {
+                    "number": 0,
+                    "percentage": 0
+                }
+            },
+            "_id": "66e62b9afaf568e0518c7fb2"
+        },
+        {
+            "CourseName": "Orthodontics I",
+            "classId": [
+                "66e60a15055142a83e52037e"
+            ],
+            "levelAverage": {
+                "Poor (Bad) Questions": {
+                    "number": 0,
+                    "percentage": 0
+                },
+                "Very Difficult Question": {
+                    "number": 0,
+                    "percentage": 0
+                },
+                "Difficult Question": {
+                    "number": 0,
+                    "percentage": 0
+                },
+                "Good Question": {
+                    "number": 0,
+                    "percentage": 0
+                },
+                "Easy Question": {
+                    "number": 0,
+                    "percentage": 0
+                },
+                "Very Easy Question": {
+                    "number": 0,
+                    "percentage": 0
+                },
+                "Total Accepted": {
+                    "number": 0,
+                    "percentage": 0
+                },
+                "Total Rejected": {
+                    "number": 0,
+                    "percentage": 0
+                }
+            },
+            "_id": "66e62b9afaf568e0518c7fb3"
+        },
+        {
+            "CourseName": "Periodontology I",
+            "classId": [
+                "66e60ac858d7ef3fc302bd87"
+            ],
+            "levelAverage": {
+                "Poor (Bad) Questions": {
+                    "number": 0,
+                    "percentage": 0
+                },
+                "Very Difficult Question": {
+                    "number": 0,
+                    "percentage": 0
+                },
+                "Difficult Question": {
+                    "number": 0,
+                    "percentage": 0
+                },
+                "Good Question": {
+                    "number": 0,
+                    "percentage": 0
+                },
+                "Easy Question": {
+                    "number": 0,
+                    "percentage": 0
+                },
+                "Very Easy Question": {
+                    "number": 0,
+                    "percentage": 0
+                },
+                "Total Accepted": {
+                    "number": 0,
+                    "percentage": 0
+                },
+                "Total Rejected": {
+                    "number": 0,
+                    "percentage": 0
+                }
+            },
+            "_id": "66e62b9afaf568e0518c7fb4"
+        },
+        {
+            "CourseName": "Dental Public Health & Community Dentistry II",
+            "classId": [
+                "66e60b5a2287e7f0cca09b2b"
+            ],
+            "levelAverage": {
+                "Poor (Bad) Questions": {
+                    "number": 0,
+                    "percentage": 0
+                },
+                "Very Difficult Question": {
+                    "number": 0,
+                    "percentage": 0
+                },
+                "Difficult Question": {
+                    "number": 0,
+                    "percentage": 0
+                },
+                "Good Question": {
+                    "number": 0,
+                    "percentage": 0
+                },
+                "Easy Question": {
+                    "number": 0,
+                    "percentage": 0
+                },
+                "Very Easy Question": {
+                    "number": 0,
+                    "percentage": 0
+                },
+                "Total Accepted": {
+                    "number": 0,
+                    "percentage": 0
+                },
+                "Total Rejected": {
+                    "number": 0,
+                    "percentage": 0
+                }
+            },
+            "_id": "66e62b9afaf568e0518c7fb5"
+        }
+    ],
+
+
+
+};
+module.exports = dummydata;
