+++ Backend/package.json
@@ -29,6 +29,7 @@
     "multer": "^1.4.5-lts.1",
     "node-cron": "^3.0.3",
     "nodemon": "^3.1.4",
+    "pdf-lib": "^1.17.1",
     "pdf-merger-js": "^5.1.2",
     "pg": "^8.12.0",
     "prisma": "^5.18.0",
